requirements.txt
playwright
google-genai
pydantic

streamlit

fastapi
uvicorn
python-multipart
playwright-stealth
